Refactoring into concerns / encapsulated trait code
