+++ src/models/mod.rs
@@ -11,6 +11,7 @@ pub use self::users::User;
 pub use self::venues::*;
 
 pub mod artists;
+pub mod concerns;
 pub mod event_artists;
 pub mod event_histories;
 pub mod events;
