+++ src/models/users.rs
@@ -1,4 +1,4 @@
-use chrono::{NaiveDateTime, Utc};
+use chrono::NaiveDateTime;
 use db::Connectable;
 use diesel;
 use diesel::prelude::*;
@@ -7,8 +7,6 @@ use schema::users;
 use utils::errors::{DatabaseError, ErrorCode};
 use utils::passwords::PasswordHash;
 use uuid::Uuid;
-
-const PASSWORD_RESET_EXPIRATION_PERIOD_IN_DAYS: i64 = 2;
 
 #[derive(Insertable)]
 #[table_name = "users"]
@@ -31,14 +29,6 @@ pub struct User {
     pub last_used: Option<NaiveDateTime>,
     pub active: bool,
     pub role: Vec<String>,
-    pub password_reset_token: Option<Uuid>,
-    pub password_reset_requested_at: Option<NaiveDateTime>,
-}
-
-#[derive(AsChangeset)]
-#[changeset_options(treat_none_as_null = "true")]
-#[table_name = "users"]
-pub struct Resetable {
     pub password_reset_token: Option<Uuid>,
     pub password_reset_requested_at: Option<NaiveDateTime>,
 }
@@ -91,16 +81,6 @@ impl User {
         )
     }
 
-    pub fn find_by_password_reset_token(password_reset_token: &Uuid, conn: &Connectable) -> Result<User, DatabaseError> {
-        DatabaseError::wrap(
-            ErrorCode::QueryError,
-            "Error loading user",
-            users::table
-                .filter(users::password_reset_token.eq(password_reset_token))
-                .first::<User>(conn.get_connection()),
-        )
-    }
-
     pub fn check_password(&self, password: &str) -> bool {
         let hash = match PasswordHash::from_str(&self.hashed_pw) {
             Ok(h) => h,
@@ -121,79 +101,8 @@ impl User {
         )
     }
 
-    pub fn create_password_reset_token(&self, conn: &Connectable) -> Result<User, DatabaseError> {
-        let data = Resetable {
-            password_reset_token: Some(Uuid::new_v4()),
-            password_reset_requested_at: Some(Utc::now().naive_utc()),
-        };
-
-        DatabaseError::wrap(
-            ErrorCode::UpdateError,
-            "Could not create token for resetting password",
-            diesel::update(self)
-                .set(data)
-                .get_result(conn.get_connection()),
-        )
-    }
-
     pub fn for_display(self) -> DisplayUser {
         self.into()
-    }
-
-    pub fn consume_password_reset_token(token: &Uuid, password: &str, conn: &Connectable) -> Result<User, DatabaseError> {
-        use schema::users::dsl::*;
-
-        let result = User::find_by_password_reset_token(token, conn);
-        match result {
-            Ok(user) => {
-                let user_id = user.id.clone();
-                let token_data : Resetable = user.into();
-                if token_data.is_expired() {
-                    Err(DatabaseError::new(
-                        ErrorCode::InternalError,
-                        Some("Password reset token is expired"),
-                    ))
-                } else {
-                    let hash = PasswordHash::generate(password, None);
-
-                    DatabaseError::wrap(
-                        ErrorCode::UpdateError,
-                        "Could not save new password for user",
-                        diesel::update(users.filter(id.eq(user_id)))
-                            .set((
-                                hashed_pw.eq(&hash.to_string()),
-                                Resetable {
-                                    password_reset_token: None,
-                                    password_reset_requested_at: None,
-                                }
-                            ))
-                            .get_result(conn.get_connection()),
-                    )
-                }
-            }
-            Err(e) => Err(e),
-        }
-    }
-}
-
-impl Resetable {
-    pub fn is_expired(&self) -> bool {
-        match self.password_reset_requested_at {
-            Some(password_reset_requested_at) => {
-                let now = Utc::now().naive_utc();
-                now.signed_duration_since(password_reset_requested_at).num_days() >= PASSWORD_RESET_EXPIRATION_PERIOD_IN_DAYS
-            }
-            None => panic!("Date of password reset not available"),
-        }
-    }
-}
-
-impl From<User> for Resetable {
-    fn from(user: User) -> Self {
-        Resetable {
-            password_reset_token: user.password_reset_token,
-            password_reset_requested_at: user.password_reset_requested_at,
-        }
     }
 }
