+++ tests/unit/concerns/users/mod.rs
@@ -1,0 +1,1 @@
+pub mod password_resetable;
